[BUGFIX] Stop requiring the static code analysis gems
Most of the static code analysis gems can be configured to require: false
in the Gemfile as they are not needed when running the App (or the specs).
The only exception is coffeelint which brings its own Rake task.

In addition, capybara can be set to require: false as well.

This commit also adds the shoulda-matchers gem.

This commit should allow switching from scss-lint to scss_lint.

Closes #97
Signed-off-by: Michael Schulze <m.schulze@braingourmets.com>